fix: move table item
